api/requirements.txt
flask==1.1
pynubank==2.10
flask-cors==3.0
pytz==2020.1
python-dateutil==2.8
pymongo==3.11
pillow==7.2